Keyword `unevaluatedProperties` (#128)
This PR addresses #83.

Some of the sentences I have added are based on my interpretation of the
results from hyperjump and json-everything. I have tried my best to
cover most of the fundamental aspects. If any sentence needs improvement
or seems incorrect, please provide a brief explanation why.

Additionally, I haven't removed the annotations section (line 26-29).
Let me know if we need to keep it or not.